+++ packages/terra-table/src/subcomponents/Row.jsx
@@ -34,6 +34,11 @@ const propTypes = {
   sectionId: PropTypes.string,
 
   /**
+   * An identifier for the subsection.
+   */
+  subsectionId: PropTypes.string,
+
+  /**
    * String that specifies height of the row. Any valid CSS width value is accepted.
   */
   height: PropTypes.string,
@@ -116,6 +121,7 @@ function Row(props) {
     id,
     tableId,
     sectionId,
+    subsectionId,
     isSelected,
     isTableStriped,
     cells,
@@ -181,6 +187,7 @@ function Row(props) {
             rowIndex={rowIndex}
             columnIndex={cellColumnIndex}
             sectionId={sectionId}
+            subsectionId={subsectionId}
             tableId={tableId}
             key={`${id}_${columnId}`}
             isSelected={!rowSelectionMode && cellData.isSelected}