+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.16.0 - (April 10, 2024)
+
+* Changed
+  * Changed to avoid focus being set to the first cell on multiple re-renders.
+
 ## 1.15.0 - (April 4, 2024)
 
 * Changed