+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.26.0 - (April 10, 2024)
+
+* Added
+  * Added `isAutoFocusEnabled` prop to allow auto focus of interactable elements when the only interactable element is a button or hyperlink.
+
 ## 1.25.0 - (April 4, 2024)
 
 * Changed