+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.80.0 - (April 10, 2024)
+
+* Added
+  * Added documentation for subsections in `terra-table`.
+  * Added examples and tests for `isAutoFocusEnabled` prop for `terra-data-grid`.
+  
 * Changed
   * Updated `date-time-picker` KeyboardShortcuts doc for +/- Keys for DateTime input.